move to folder component
