+++ nodes/worker/app.py
@@ -3,6 +3,6 @@ import utils
 from time import sleep
 from threading import Thread
 from enum import Enum
-from . import availibility, cancel, executor, status
+from component import availibility, cancel, executor, status
 
 # TODO bikin thread masing-masing port
